requirements.txt
crewai==0.19.0
anthropic==0.18.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
trafilatura==1.6.1